Add Bootstrap-Vue.  Weird Cytoscape canvas display issue
Cytoscape's canvas is only using the right-hand 50% of its container.
The canvas elements are getting placed with their left edges
horizontally centered within the #cy div element.  Click events are
being offset as if the canvases were placed at the absolute left side
of the container.
Everything behaves as expected if I manually update the style of the
canvas elements to have the property "left: 0px".  But this is a hack
and not an appropriate solution in the long run, I think.