Build working jar with all dependencies, including client
Run the following command in the directory "server":
mvn clean install && mvn assembly:single && java -jar target/web-frontend-1.0-SNAPSHOT-jar-with-dependencies.jar

Then open localhost:4567 in your browser.

TODO: Remove hardcoded reference to "localhost:4567" in client-side code.
(The client is supposed to eventually run on a PC, while the server runs
on a University server.)
Set things up so that "npm run dev" will still be usable.

"npm run dev" puts up its own web server, so it may be appropriate to
have a switch: In production mode, the client will use a relative
URL to reach the server (since it can be assumed that the client will
be served from the same web server as the REST API),
while in development mode, the client could use a hard coded
localhost URL for the REST API, exactly as it does now.