Refactor: Construct Json automatically using Gson.toJsonTree(Object)
This significantly lessens the boilerplate needed to implement our
AJAX interface.

It's still fairly brittle, like any interface implemented this way.
I need to implement error handling and validate data on both the server
and client sides. Despite the drawbacks, I chose to do it this way
because APIs using JSON and AJAX are fairly widespread, so if I leave the
project, it's likely that another student will be able to pick it up in
my place.

I also considered using Scala.js for this project. It's possible in Scala.js
to easily define type-checked Ajax calls using the Autowire library;
see this blog post by Li Haoyi, a core developer of Scala.js:
http://www.lihaoyi.com/post/FromfirstprinciplesWhyIbetonScalajs.html#type-checked-ajax-calls
I find this to be a significant point in favor of Scala.js.  That said,
I think it would be harder to find another student to maintain a
project written in Scala.