Bug fix: You no longer have to press "get strategy BDD" twice
This fixes the bug where you would have to click "Get Strategy BDD"
twice before the strategy BDD would actually appear in the editor.

I fixed this by initializing the "nodes" and "links" arrays in the
GraphEditor component in a more sane way.

Before, I was using a hardcoded graph from an example I found online.

Now, this component will use, from the very beginning, the graph that
is provided to it via the "petriNet" property.