Bug fix: Multiple graph editor instances can now coexist
The problem was that the SVG elements of all GraphEditor instances
had the same ID, so all the graphs would all end up on top of each
other.

Now, we give a unique ID to each SVG per instance.
This way, multiple instances of the GraphEditor are able to coexist.
They won't all put their graphics into the same SVG any more; each
instance of the component finds its own SVG by using its unique ID.

I also changed some other css rules to use classes instead of ids.
I think the rule is, when designing a component, if you expect to
have more than one instance of it, then all of the ids used in it should
incorporate the instance's this._uid attribute.  Otherwise, you end up
with duplicate IDs, which is incorrect -- DOM element IDs are supposed
to be unique.