Show Petri Game and Strategy BDD in two different "tabs"
TODO: Use some kind of event listener in GraphEditor so that the
centering force, which keeps the graph in the center of the editor,
will get updated when the component is made visible.

E.g. the DOM MutationEvent interface.
https://www.w3.org/TR/DOM-Level-2-Events/events.html#Events-eventgroupings-mutationevents

(Currently, the graph will sort of hang out in the upper left corner
of the SVG until we hear a window resize event.

I think happens because, at the time that the center force is
initialized (i.e. at the time the GraphEditor component is mounted),
the component is invisible and the SVG basically counts as one pixel.