Throw out strategyBDD when it becomes obsolete
Automatically change tabs between Petri Game/Strategy BDD
when appropriate.

To make tab-switching work, disable Vue's Router (which we
weren't using).
(Tabs are switched in vue-tabs-component using the #fragment at the end
of a URL.  Normally, Vue's Router makes use of the fragment, so the two
extensions seem to conflict with one another.)