Display result of existsWinningStrategy in toast notification
I'm not sure this is the best way to display the information.  It should
probably also be available in some more persistent way so that the user
would not have to click the button again to be reminded of the result.

I think this piece of state should be kept track of on the server.
Same with the strategy BDD for a given PetriGame.  See my notes
in App.vue.