Bug fix: Display arrowheads in multiple instances of Petri Net viewer
The problem was that arrowheads were only getting shown for the Petri
Game, but not for the strategy BDD.

The solution was to give the arrowhead definition a unique ID for each
SVG on the page.

Also, add some comments about how I plan to fix the positions of the
graphs so that they don't always start off in the upper left corner.
This also affects our performance.  The FPS feels noticably lower to me
after I started using the IntersectionObserver API polyfill to solve
this issue.