Display graph strategy BDD in a tab (lacking visual style though)
I'm following the same pattern I used to display the strategy BDD
so that I can quickly get this feature working.
There is some duplicated code that could bear to be refactored.

Also, like I wrote in my comments in App.vue, there is a race condition
when the user clicks "get graph strategy BDD" and "send graph to
editor" in quick succession.  For this I have a solution that is also
described briefly in the comments.

Also, you can crash the server by clicking "get graph strategy BDD"
a bunch of times really fast.  This will also be addressed in the
course of fixing the race condition that I mentioned.