Bug fix: Change nodes in graph viewer to reflect changes in APT
This fixes the bug where, if you had created a Petri Game and then
edited the APT so a System place became an Environment place, the
change would not be reflected in the graph viewer.

Another symptom of this bug was that, if you deleted a Place and then
created a Transition with the same ID, the Transition would be drawn
as a circle (like a Place) rather than a square like it's supposed to.

I fixed it by changing the key function for our nodes.
(i.e. The second argument to D3's .data() function.)

The key is now made up of not just the node's ID, but also its type
(i.e. ENVPLACE vs. SYSPLACE vs. TRANSITION ...).  This way, if the type
of a node changes, D3 will delete the old, no-longer-needed element
when we call nodeElements.exit().remove().