Calculate Graph Game BDD and display it in the browser
It's pretty slow.  Even after you update the Petri Game so that the
Graph Game BDD disappears, the client remains sluggish.  I think this
happens because the force simulation for the deleted Graph Game
is still happening in the background.  Refreshing the page solves this
for now, but in the future I should stop the simulation so that no
slowdown will occur.