wip refactor 3.
TODO: Reconsider encapsulation of PetriNetD3 class.
I thought it would be smart to prevent us from ever getting
ahold of a PetriNetD3 object, since in theory it is useless except
for being JSONified.  But then I realized I was nesting the PetriNetD3
inside of PetriGameD3 in order to attach a UUID to it.
So the question now is, should I keep doing things this way or should
I find a difefrent way to do it