re-enable /getGraphStrategyBDD.
Add TODO note about issue that can crash the server if you click
e.g. existsWinningStrategy many times very fast.
This isn't a problem if you give it a second after the first time you
click it.  At that point the answer is cached and will not be
re-calculated.  But if you start with a fresh Petri Game and hammer
on the button right from the start, you can cause the answer to be
calculated many times in parallel, using up all the JVM's memory.