Expand and collapse Graph Game states by clicking on them
The client UI is still a bit weird.  The graph does not appear to be
updated in place, but rather completely reset (popping back up to the
upper-left corner of the screen) each time you expand/collapse a node.

There is a fix for this.  I wrote some TODO comments about it.
There is also kind of weird behavior when a child of a node has had
its own children be expanded, and then you collapse the original
grandparent node.  I also wrote a TODO about that.