Improve display of BDDGraphs
The color of each State is set to gray or white depending on isMcut.
The border is a thick, green dashed line for "good" states
The border is a thick black line for "bad" states

In the "Graph Game BDD" viewer, the IDs of states are bold if they
have not yet been clicked on to show their children.
(This was previously indicated through the color of the states.)