GraphEditor: Clicking to expand States works better
It used to be that you had to click perfectly in order to show a State's
postset.  If you dragged the State even a tiny bit by mistake, your
left click wouldn't be registered.

Now, there is a dead zone built into clicking and dragging.
If you drag a point less than a certain distance, then your input will
get registered as a click.  This should be much less frustrating.