WIP, draw asterisks by StrategyBDD when Petri Game is modified
Right now the one problem is, there is no message that gets shown to
the user to explain what the asterisks mean.  I want there to be a
pop-up alert saying "The Petri Game in the editor does not match the
one that this Strategy BDD is based on.  Do you want to recalculate
the strategy BDD?  yes/no".  But I don't see an obvious way to do this
using vue-tabs-component.  I'm thinking that I will just stop using it
and put the tabs inside of a toolbar at the top of the screen