Implement WebSocket for logging (wip)
This is just a proof of concept.
Note: Right now, if multiple users are connected at once, they will all
see the output of each other's actions on the server.  Nothing is being
done yet to split up the output per user.  There might be race
conditions...
e.g. If you try to take over ADAM's output at the start of handling a
request, then what happens if two requests occur simultaneously?
I'm not sure how ADAM handles this but I'm pretty sure it is happy to
be spun up in multiple threads at once, each sharing this global output
stream variable.  So my thought is to either
a) create a queue of ADAM requests and run them one at a time, (which is
not a super awesome solution in case you really have multiple users)
or
b) Change ADAM so it doesn't have this global state, rather calling
Adam.createInstance() or something at the start of a program.

idk.  just brainstorming ideas.