Tweak graph editor height.  Not a real fix.
The issue here is twofold:

1) Graph editor does not grow to fill the space available to it.
It has a hardcoded "80vh" height set to it inside of itself.

2) The graph editor doesn't shrink when the log viewer is opened.
Instead, the log viewer just opens up below the graph viewer,
making the page stretch downwards.
I want the page to stay 100% of window height and have the log viewer
push up the bottom edge of the graph viewer.  how can I do this with
Vuetify?  I don't want to hand-code something like I had before
if I can avoid it.

I guess a possible solution to #1 may involve giving the desired height
in px to the GraphEditor component as a property.
As for #2, I have a hunch that hand coding is the only way around this.
Vuetify does not seem particularly flexible.  Actually, most frameworks
seem to just throw up their hands and shrug when it comes to defining
any kind of UI that is not a cookie cutter copy of some other website.