Bug fix: Don't delete x/y positions upon parsing error
I used to throw out the petri game if there was a parsing error.
I don't do that any more.
You can edit the APT in peace without worrying that a single typo or
moment of hesitation mid-line will result in your carefully laid out
nodes and transitions being scattered into the wind.

The parsing error will be printed in the log at the bottom of the page.