Automatically scroll log viewer to bottom when changing settings
An even better behavior would be to stay at the same time in the log.
This would require figuring out what "time" the log is currently
scrolled to and being able to scroll to a given time.
That is in principle doable, but will take a little bit more work, so
I am leaving it how it is for now.