Re-enable zoom
Note that at the moment, nodes you move apparently may have their
positions reset when a new node is inserted.  The problem is if you
have a position saved in the APT on the backend, and then you
 move a node without saving its position to the backend,
and then you insert a new node, you get a copy of the JSON reflecting
the old x/y coordinates from the backend, and these are applied to
the graph in the clietn gui editor, resulting in the nodes unexpectedly
snapping back to their old positions.
I will have to think about the best way to deal with this.