Bug fix: "Zoom to fit all nodes" zooms really far out
I had a feature in the works that w'yud add a
border around a selection of nodes, and you should be able to click
and drag the border to stretch and shrink the selection.

I have been drawing the border all the time, it's just been invisible
to the user.
Sometimes it would get messed up and be really huge.
In that case, "zoom to fit all nodes" would zoom out far enough to fit
the very big, invisible border.  So it would zoom REALLY far out!
Now it won't do that any more.  :)