WIP draw token flow
a source, transition, and postset can be gathered by clicking on
them in succession.  pressing "enter" sends an event that should trigger
a request to the server to create the given token flow.

todo: send the request and process it on the server
todo: reset the state of the tool when switching to/from it
todo: visualize the selection while it is in progress