Add button to finish drawing token flow (workaround for linux?)
I'm not sure if this is the case on my computer in the office, but in
my virtualbox vm, I can't seem to get any keyboard events to register.
To work around this issue, I added a button that does the same thing
as if you would press Enter.

I also added "Return" as a synonym for the "Enter" key.  I seem to
recall that this was necessary on my office PC when I was testing
keyboard input before.  That seems to indicate to me that keyboard
input works at least on some computers running Linux.  It might just be
a quirk of my own setup in Virtualbox.  Just to be safe,
I'm adding this button to the toolbar anyway.