Send updated PetriGame to client when partition information may have been added
I am pretty sure there is a better way to do this.  For example, the
server could push a message to the client over a WebSocket whenever the
PetriGame's state may have changed.
For now, I'm just doing it the 'easiest' way I can think of without
doing a bunch of refactoring.