Add console.log statement for debugging the physics simulation with vue-router
The issue with the approach using vue-router to switch between GUIs
is that, if the user uses the browser's back button to go back to the
home page and open the other GUI, the physics simulations in the first
GUI will keep running, and this is actually a fairly CPU intensive thing
that can cause the framerate to drop on slower computers.

My plan is to find a way in to check with each simulation
tick if the GraphEditor is visible and sleep for a bit if it doesn't
need to be running.