If Graph Game BDD is calculated in 5 seconds or less, it gets loaded right away
Appropriate messages are shown in the log and in the message box thing
at the bottom of the screen:

"Sent request to server"
and then
"The calculation finished" /
"It's taking more than 5 seconds.  It will run in the background."

TODO: Distinguish between whether it has merely been queued up or if it is actually running.
Finish the GUI for showing hte list of running jobs.