Allow queuing up model checking jobs and showing them in the list.  Rough edges.
It's not really production ready because it doesn't differentiate
between jobs to check different formulas when determining whether a job
has already been run or not.  That is, once you check one formula,
you have to go into the calculation list and delete the previous job
before you are allowed to queue up another one.

Also, the job list does not show what LTL formula was actually checked
by a given job, and it just says "FALSE" "TRUE" or "UNKNOWN" instead
of a nice human-readable text.