Prepare to fire transitions in either editor nets or MC nets/winning strategies
The two are stored in different places (in a central Map in App, vs.
in a client's own UserContext object), so I have added the possibility to
supply a 'getter' function to find a PetriNet given the parameters of
the user's request, wherever that PetriNet is supposed to be stored.