Fix #25 by getting rid of ping system
This bug arose because the server was calling
"RemoteEndpoint::sendString(String text)" from multiple threads,
which you're not supposed to do.
https://bugs.eclipse.org/bugs/show_bug.cgi?id=474488

In contrary to what the comment in the code says,
you don't have to send ping messages to keep a websocket alive,
so this "ping" / "pong" feature I made is not needed.
For the purpose of verifying that the Websocket connection is alive,
you can open DevTools in the browser in FF/Chrome and check it there.