Refactor 'updateLinkStroke' to 'updateLinkEl'
This facilitates breaking out and reusing a section of code from
updateD3(), making that very long function a little bit shorter.

The performance hit compared to the previous commit seems to be minimal.
With intensive mousing-over of many nodes and links, CPU usage does not
reach over 40% on my machine, and there is no perceptible rendering lag.