Bug fix: Jobs should operate upon copies of editor nets
It was until now possible to modify the net in the editor while a job
was running using the same underlying PNWT/PG object.  This may have
caused a variety of interesting and hard to reproduce bugs.

I forsaw this and went through the trouble of making a copy of every
net before queueing up each job to prevent this kind of thing from
happening, but unfortunately, I made a small typo and passed the
original net, rather than the copy, to the method 'makeJob'.